+++ legacy/tsconfig.json
@@ -7,7 +7,7 @@
 			"@test/*": ["./test/*"]
 		},
 		"rootDir": ".",
-		"types": ["jest", "node"]
+		"types": ["node"]
 	},
 	"include": ["src/**/*.ts", "test/**/*.ts", "scripts/**/*.ts", "tsup-types.config.ts"],
 	"exclude": ["node_modules", "dist"]
