+++ package.json
@@ -8,7 +8,13 @@
 	"main": "index.js",
 	"scripts": {
 		"build": "turbo build",
-		"build:core": "turbo build --filter=@tmlmobilidade/core",
+		"build:connectors": "turbo build --filter=@tmlmobilidade/connectors",
+		"build:interfaces": "turbo build --filter=@tmlmobilidade/interfaces",
+		"build:lib": "turbo build --filter=@tmlmobilidade/lib",
+		"build:providers": "turbo build --filter=@tmlmobilidade/providers",
+		"build:types": "turbo build --filter=@tmlmobilidade/types",
+		"build:ui": "turbo build --filter=@tmlmobilidade/ui",
+		"build:utils": "turbo build --filter=@tmlmobilidade/utils",
 		"dev": "turbo dev",
 		"lint": "turbo lint",
 		"clean": "sh cleanup.sh",
