+++ packages/connectors/tsconfig.json
@@ -6,7 +6,7 @@
 			"@/*": ["./src/*"]
 		},
 		"rootDir": ".",
-		"types": ["jest", "node"]
+		"types": ["node"]
 	},
 	"include": ["src/**/*.ts", "test/**/*.ts", "scripts/**/*.ts"],
 	"exclude": ["node_modules", "dist"]
