+++ packages/types/package.json
@@ -31,16 +31,19 @@
 		"lint:fix": "eslint --fix"
 	},
 	"exports": {
+		"default": "./dist/src/index.js",
 		"types": "./dist/src/index.d.ts"
 	},
 	"dependencies": {
+		"geojson": "0.5.0",
+		"luxon": "3.5.0",
+		"mongodb": "6.15.0",
 		"zod": "3.24.2"
 	},
 	"devDependencies": {
 		"@carrismetropolitana/eslint": "20250303.2009.45",
 		"@dotenv-run/cli": "1.3.6",
 		"@tmlmobilidade/tsconfig": "*",
-		"@tmlmobilidade/utils": "*",
 		"@types/luxon": "3.4.2",
 		"@types/node": "22.13.13",
 		"resolve-tspaths": "0.8.23",
