+++ packages/types/src/ride.ts
@@ -1,7 +1,6 @@
 /* * */
 
-import { DocumentSchema, type OperationalDate, type UnixTimestamp } from '@/common.js';
-import { validateOperationalDate, validateUnixTimestamp } from '@tmlmobilidade/utils';
+import { DocumentSchema, type OperationalDate, type UnixTimestamp, validateOperationalDate, validateUnixTimestamp } from '@/common.js';
 import { z } from 'zod';
 
 /* * */
