+++ packages/utils/package.json
@@ -50,8 +50,9 @@
 	"devDependencies": {
 		"@carrismetropolitana/eslint": "20250303.2009.45",
 		"@tmlmobilidade/tsconfig": "*",
+		"@tmlmobilidade/types": "*",
 		"@types/luxon": "3.4.2",
-		"@types/node": "22.13.10",
+		"@types/node": "22.13.13",
 		"resolve-tspaths": "0.8.23",
 		"rimraf": "6.0.1",
 		"turbo": "2.4.4",
