+++ packages/utils/src/operational-date.ts
@@ -1,19 +1,7 @@
 /* * */
 
-import { OPERATIONAL_DATE_FORMAT, type OperationalDate } from '@/types/index.js';
+import { type OperationalDate } from '@tmlmobilidade/types';
 import { DateTime } from 'luxon';
-
-/**
- * This function validates if a string is a valid operational date.
- * Throws an error if the date is invalid.
- * @param date - The date to be validated.
- * @returns The given string as an OperationalDate.
- */
-export function validateOperationalDate(date: string): OperationalDate {
-	const parsedDate = DateTime.fromFormat(date, OPERATIONAL_DATE_FORMAT);
-	if (!parsedDate.isValid) throw new Error(`Invalid date format '${date}', expected format: ${OPERATIONAL_DATE_FORMAT}, explanation: ${parsedDate.invalidExplanation}`);
-	return parsedDate.toFormat(OPERATIONAL_DATE_FORMAT) as OperationalDate;
-}
 
 /**
  * Returns the operational date based on the provided timestamp and format.
